Rename index
Personal Notes [not formatted for audience]:
Renaming index.html, back to original [local] name "Hemingways - Marietta, Georgia.html".  Purpose:  to test if this is a factor after 4 attempts at pushing the files here and the entire tour not loading [specifically with error 'the loading of "xyz" .jpg failed].
Locally checked the files for corruption.  Tour was generated in Panotour and I rebuilt the project in that program into a new DIR and filename;  plus replaced the pano pics within the project with the same ones [thinking it could be a 'renaming' of a parent dir issue]- but issue still persists.
Previous upload pushes have been done in the same exact way, and tours display and function 100%.  
Renaming 'index' file now, to see if the XMLs are looking for a particular filename but the index file is named 'index' [which was also a renaming before uploading and after producing and finishing the tour].  Again, this is how the other tours have been committed and pushed, and the original HTML file [locally] would be named 'ABCXYZ' for example, and it would be renamed to 'index', then pushed, and worked fine.